+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Added
 - `Spinner` widget that displays a looping animation
 - `EmptyEidget` widget that can be used as a placeholder in layouts
+- `row{}` widget layout that create a single horizontal row of widgets
 
 ### Fixed
 - Reduced flickering on high frame rate animations